Update README.md
feet/inches to meters

Write a program that reads in a length in feet and inches and outputs the equivalent length in meters. Create and use these functions for the subtasks:

one for input (with 2 reference parameters)
one for calculating
one for displaying result
There are 0.3048 meters in a foot and 12 inches in a foot. Here is example code to do the conversion:
meters = METERS_PER_FOOT * (feet + inches/INCHES_PER_FOOT);

Optional:
Include a loop that lets the user repeat this computation for new input values until the user says they want to end the program.